Replace shouldSkipRequest() with FilterChain implementation
